Rename OpenAL audiochannel
